support multiple G72
