+++ src/test/java/com/walmartlabs/x12/dex/dx894/DefaultDex894ParserTest.java
@@ -15,7 +15,6 @@ limitations under the License.
  */
 package com.walmartlabs.x12.dex.dx894;
 
-import com.walmartlabs.x12.exceptions.X12ErrorDetail;
 import com.walmartlabs.x12.exceptions.X12ParserException;
 import org.junit.Before;
 import org.junit.Test;
@@ -23,13 +22,9 @@ import org.junit.Test;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.List;
-import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
 
 public class DefaultDex894ParserTest {
 
@@ -119,110 +114,6 @@ public class DefaultDex894ParserTest {
     public void testParsingInvalidSegments() throws IOException {
         byte[] dexBytes = Files.readAllBytes(Paths.get("src/test/resources/dex/894/dex.sample.invalid.segments.txt"));
         dexParser.parse(new String(dexBytes));
-    }
-
-    /**
-     * TODO: G72 needs to be figured out
-     */
-    @Test
-    public void testParsingMultipleG72() throws IOException {
-        byte[] dexBytes = Files.readAllBytes(Paths.get("src/test/resources/dex/894/dex.sample.multiple.g72.txt"));
-        Dex894 dex = dexParser.parse(new String(dexBytes));
-        assertNotNull(dex);
-
-        DefaultDex894Validator dexValidator = new DefaultDex894Validator();
-        Set<X12ErrorDetail> errors = dexValidator.validate(dex, false);
-        assertNotNull(errors);
-        assertEquals(0, errors.size());
-
-        // DXS segment
-        assertEquals("9269850000", dex.getSenderCommId());
-        assertEquals("DX", dex.getFunctionalId());
-        assertEquals("004010UCS", dex.getVersion());
-        assertEquals("3", dex.getHeaderTransmissionControlNumber());
-        assertEquals("9254850000", dex.getReceiverCommId());
-        assertEquals("P", dex.getTestIndicator());
-
-        // DEX transactions
-        List<Dex894TransactionSet> dexTxList = dex.getTransactions();
-        assertNotNull(dexTxList);
-        assertEquals(1, dexTxList.size());
-
-        //
-        // DEX transaction (1)
-        //
-        Dex894TransactionSet dexTx = dexTxList.get(0);
-        assertNotNull(dexTx);
-        // ST segment
-        assertEquals("894", dexTx.getTransactionSetIdentifierCode());
-        assertEquals("0002", dexTx.getHeaderControlNumber());
-        // SE segment
-        assertEquals("0002", dexTx.getTrailerControlNumber());
-        assertEquals(new Integer(13), dexTx.getExpectedNumberOfSegments());
-        assertEquals(new Integer(13), dexTx.getActualNumberOfSegments());
-        // G82 segment
-        assertEquals(InvoiceType.D, dexTx.getDebitCreditFlag());
-        assertEquals("109152000116", dexTx.getSupplierNumber());
-        assertEquals("051957769", dexTx.getReceiverDuns());
-        assertEquals("0001", dexTx.getReceiverLocation());
-        assertEquals("066563859", dexTx.getSupplierDuns());
-        assertEquals("000000", dexTx.getSupplierLocation());
-        assertEquals("20181129", dexTx.getTransactionDate());
-        // G84 segment
-        assertEquals("340.000", dexTx.getTransactionTotalQuantity().toString());
-        assertEquals("120931.00", dexTx.getTransactionTotalAmount().toString());
-        assertEquals(null, dexTx.getTransactionTotalDepositAmount());
-        // G85 segment
-        assertEquals("XXXX", dexTx.getIntegrityCheckValue());
-        // G86 segment
-        assertEquals("7898", dexTx.getElectronicSignature());
-        assertEquals(null, dexTx.getSignatureName());
-        // transaction data
-        assertNotNull(dexTx.getTransactionData());
-        assertTrue(dexTx.getTransactionData().startsWith("ST*894*0002"));
-        assertTrue(dexTx.getTransactionData().contains("G83*25*12*EA*007192167217*AC*TMB ORG 12IN 12CT**3.8**OR PEP 12\r\n"));
-        assertTrue(dexTx.getTransactionData().endsWith("G86*7898\r\n"));
-
-        // items
-        List<Dex894Item> dexItemList = dexTx.getItems();
-        assertNotNull(dexItemList);
-        assertEquals(2, dexItemList.size());
-
-        // DEX transaction (1) item (1)
-        Dex894Item dexItem = dexItemList.get(0);
-        assertNotNull(dexItem);
-        // G83 segment
-        assertEquals("25", dexItem.getItemSequenceNumber());
-        assertEquals("12.000", dexItem.getQuantity().toString());
-        assertEquals(UnitMeasure.EA, dexItem.getUom());
-        assertEquals("007192167217", dexItem.getUpc());
-        assertEquals(ProductQualifier.UNKNOWN, dexItem.getConsumerProductQualifier());
-        assertEquals("TMB ORG 12IN 12CT", dexItem.getConsumerProductId());
-        assertEquals(null, dexItem.getCaseUpc());
-        assertEquals("3.80", dexItem.getItemListCost().toString());
-        assertEquals(null, dexItem.getPackCount());
-        assertEquals("OR PEP 12", dexItem.getItemDescription());
-        assertEquals(null, dexItem.getCaseProductQualifier());
-        assertEquals(null, dexItem.getCaseProductId());
-        assertEquals(null, dexItem.getInnerPackCount());
-
-        // DEX transaction (1) item (2)
-        dexItem = dexItemList.get(1);
-        assertNotNull(dexItem);
-        // G83 segment
-        assertEquals("27", dexItem.getItemSequenceNumber());
-        assertEquals("24.000", dexItem.getQuantity().toString());
-        assertEquals(UnitMeasure.EA, dexItem.getUom());
-        assertEquals("007465339393", dexItem.getUpc());
-        assertEquals(ProductQualifier.UNKNOWN, dexItem.getConsumerProductQualifier());
-        assertEquals("JCK PZA 12IN 12CT", dexItem.getConsumerProductId());
-        assertEquals(null, dexItem.getCaseUpc());
-        assertEquals("2.80", dexItem.getItemListCost().toString());
-        assertEquals(null, dexItem.getPackCount());
-        assertEquals("TC PEPSAUS 12", dexItem.getItemDescription());
-        assertEquals(null, dexItem.getCaseProductQualifier());
-        assertEquals(null, dexItem.getCaseProductId());
-        assertEquals(null, dexItem.getInnerPackCount());
     }
 
 }
